feat(pipeline): new aatams_sattag_qc_dm pipeline
Included here is:
    1. CSVSchema class
    2. AATAMS_DM_Schema class
    3. AATAMS_DM pipeline handler

The schema classes are "LEVT" classes. They can Load, Extract, Validate and Transform the required csv files into dictionaries/lists.
In particular, the AATAMS class can do schema validation and transformations in several levels:

    a.filename convention in the zip
    b.headers of all filenames
    c.types of all columns
    d.type-cast all columns
    e.stores the type-cast of the files.

The nice thing about the implementation is that, after the LEVT flow is done, we could potentially just load the csv (or the dicts/lists) into the database directly (as long as the dbschema is defined and we got authentications).

A full content validation is done by default, but that can be easily changed.

Tests:
1. Schema is validated against an aggregation of all the zip files I found in the error directory.
2. Pipeline tests are limited but we need to discuss where the err files are defacto invalid or valid.

Performance: I aggregated all aatams_dm csv in the error folder and got about ~370mb. Reading this took a bit but not much with the current implementation(<5min). It may be an issue if the files are too big. This is the reason some options are provided that may skip some rows in the content validation.