feat(aatams): block non-unique campaign files
This is a fix-up to be even stricter regarding
the file name conventions, campaign id's,
and csv content.

This commit will only allow ingestion if
the campaign id within the zip filename,
and csv filenames, is (are) the only campaign(s)
included within all csvs content.

Before, we only checked for campaign consistency
among all csvs content.

Now, the campaign filename id must match the csv
content campaign of all files. This is consistent
with what the harvester is doing.